Fix #1328 Missing notifications: initializeIncomingMessageNotifier() in ApplicationDcContext, not ApplicationContext
